fix(SystemMetricsWidgetProvider): cast return value to int for account age calculation
- Updated the return statement in getAccountAgeDays method to explicitly cast the result of diffInDays to int, ensuring type consistency.
- This change improves the reliability of the account age calculation in the SystemMetricsWidgetProvider.

+++ app/BusinessModules/Features/AdvancedDashboard/Services/Widgets/Providers/Performance/SystemMetricsWidgetProvider.php
@@ -106,6 +106,6 @@ class SystemMetricsWidgetProvider extends AbstractWidgetProvider
             return 0;
         }
 
-        return Carbon::parse($created)->diffInDays(Carbon::now());
+        return (int) Carbon::parse($created)->diffInDays(Carbon::now());
     }
 }
